rename `StateHash` to `ParityChecker`
hash is an implementation detail, and it makes it sound like a hash
container type
