+++ src/Core/CMakeLists.txt
@@ -26,7 +26,7 @@ add_library(Mupen64RR.Core STATIC
     "Memory/Memory.hpp"
     "Memory/Pif.hpp"
     "Memory/Savestates.hpp"
-    "Memory/StateHash.hpp"
+    "Memory/ParityChecker.hpp"
     "Memory/Summercart.hpp"
     "Memory/TLB.hpp"
     "R4300/Debug.hpp"
@@ -56,7 +56,7 @@ add_library(Mupen64RR.Core STATIC
     "Memory/Memory.cpp"
     "Memory/Pif.cpp"
     "Memory/Savestates.cpp"
-    "Memory/StateHash.cpp"
+    "Memory/ParityChecker.cpp"
     "Memory/Summercart.cpp"
     "Memory/TLB.cpp"
     "R4300/Debug.cpp"
