+++ src/Core/Memory/ParityChecker.hpp
@@ -17,27 +17,27 @@
  * The comparison itself is external: run the same movie on two builds with hashing enabled and diff the logged
  * hashes. Hooked from the VCR playback tick and finalized at movie end; driven by the `--state-hash` CLI flag.
  */
-namespace StateHash
+namespace ParityChecker
 {
-    /**
-     * \brief Begins a hashing run, resetting the running hash and checkpoints.
-     * \param interval Sample interval between checkpoints. Values below 1 are clamped to 1.
-     */
-    void begin(int32_t interval);
+/**
+ * \brief Begins a hashing run, resetting the running hash and checkpoints.
+ * \param interval Sample interval between checkpoints. Values below 1 are clamped to 1.
+ */
+void begin(int32_t interval);
 
-    /**
-     * \brief Records a checkpoint if the current sample lands on the interval. No-op when inactive.
-     * \param sample The current movie sample.
-     */
-    void on_sample(int32_t sample);
+/**
+ * \brief Records a checkpoint if the current sample lands on the interval. No-op when inactive.
+ * \param sample The current movie sample.
+ */
+void on_sample(int32_t sample);
 
-    /**
-     * \brief Finalizes the run and logs the final hash and per-checkpoint hashes. No-op when inactive.
-     */
-    void end();
+/**
+ * \brief Finalizes the run and logs the final hash and per-checkpoint hashes. No-op when inactive.
+ */
+void end();
 
-    /**
-     * \brief Gets whether a hashing run is currently active.
-     */
-    bool active();
-}
+/**
+ * \brief Gets whether a hashing run is currently active.
+ */
+bool active();
+} // namespace ParityChecker
